requirements.txt
pdfminer.six==20221105
unstructured==0.7.6
fastapi==0.95.2
python-multipart==0.0.6
uvicorn==0.22.0